+++ src/main/java/com/example/LibraryAPI/dto/BookResponse.java
@@ -4,13 +4,13 @@ import io.swagger.v3.oas.annotations.media.Schema;
 
 @Schema(description = "Book response")
 public class BookResponse {
-    private int id;
+    private Long id;
     private String title;
     private String author;
     private String isbn;
     private int publishedYear;
 
-    public BookResponse(int id, String title, String author, String isbn, int publishedYear){
+    public BookResponse(Long id, String title, String author, String isbn, int publishedYear){
         this.id = id;
         this.title = title;
         this.author = author;
@@ -18,7 +18,7 @@ public class BookResponse {
         this.publishedYear = publishedYear;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
