+++ src/main/java/com/example/LibraryAPI/repository/LoanRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface LoanRepository extends JpaRepository<Loan, Integer> {
-    boolean existsByBookIdAndReturnDateIsNull(int bookId);
+public interface LoanRepository extends JpaRepository<Loan, Long> {
+    boolean existsByBookIdAndReturnDateIsNull(Long bookId);
     List<Loan> findByReturnDateIsNull();
 }
