+++ src/main/java/com/example/LibraryAPI/service/BookService.java
@@ -29,7 +29,7 @@ public class BookService {
         return bookRepository.findAll();
     }
 
-    public Book findById(int id) {
+    public Book findById(Long id) {
         return bookRepository.findById(id).orElseThrow(() -> new BookNotFoundException(id));
     }
